modificación del archivo crud_persona

+++ crud_persona.py
@@ -8,17 +8,22 @@ def crear_persona(nombre, edad, correo):
 
 def modificar_persona(persona_id,nuevo_nombre, nueva_edad, nuevo_correo):
     persona=session.get(Persona,persona_id)
-    persona.nombre=nuevo_nombre
-    persona.edad=nueva_edad
-    persona.correo_electronico=nuevo_correo
-    session.commit()
+    if persona:
+        persona.nombre=nuevo_nombre
+        persona.edad=nueva_edad
+        persona.correo_electronico=nuevo_correo
+        session.commit()
+    else:
+        print(f"No se encuentra la persona con ID {persona_id}")
 
 
 def ver_personas():
-    for persona in session.query(Persona).all():
-        print(persona.id,persona.nombre, persona.edad, persona.correo_electronico)
+    return session.query(Persona).all()
 
 def eliminar_persona(persona_id):
     persona=session.get(Persona,persona_id)
-    session.delete(persona)
-    session.commit()
+    if persona:
+        session.delete(persona)
+        session.commit()
+    else:
+        print(f"No se encuentra la persona con ID {persona_id}")
